MULE-18417: Connector test failing with java.lang.IllegalStateException at LazyMuleArtifactContext.initializeComponents
