+++ integration/src/test/java/org/mule/test/integration/locator/LazyInitConfigurationComponentLocatorTestCase.java
@@ -55,7 +55,6 @@ import org.junit.Test;
 
 import io.qameta.allure.Description;
 import io.qameta.allure.Feature;
-import io.qameta.allure.Issue;
 import io.qameta.allure.Story;
 
 @Feature(CONFIGURATION_COMPONENT_LOCATOR)
@@ -74,7 +73,7 @@ public class LazyInitConfigurationComponentLocatorTestCase extends AbstractInteg
   public SystemProperty path = new SystemProperty("path", "path");
 
 
-  private static final int TOTAL_NUMBER_OF_LOCATIONS = 154;
+  private static final int TOTAL_NUMBER_OF_LOCATIONS = 116;
   @Inject
   private Registry registry;
 
@@ -94,10 +93,7 @@ public class LazyInitConfigurationComponentLocatorTestCase extends AbstractInteg
         "org/mule/test/integration/locator/component-locator-levels-config.xml",
         "org/mule/test/integration/locator/component-locator-os-connector.xml",
         "org/mule/test/integration/locator/component-locator-spring-config.xml",
-        "org/mule/test/integration/locator/component-locator-reference-component-models.xml",
-        "org/mule/test/integration/locator/module-with-config-oauth.xml",
-        "org/mule/test/integration/locator/module-with-config-http-oauth-auth-code.xml",
-        "org/mule/test/integration/locator/module-with-config-http-noconfig.xml"};
+        "org/mule/test/integration/locator/component-locator-reference-component-models.xml"};
   }
 
   @Override
@@ -316,43 +312,6 @@ public class LazyInitConfigurationComponentLocatorTestCase extends AbstractInteg
                                   "justAnotherFlowThatShouldNotBeInitialized",
                                   "justAnotherFlowThatShouldNotBeInitialized/processors/0",
 
-                                  "tokenManagerConfig-sample-config",
-                                  "github-httpreq-config-sample-config",
-                                  "github-httpreq-config-sample-config/connection",
-                                  "github-httpreq-config-sample-config/connection/0",
-                                  "github-httpreq-config-sample-config/connection/0/0",
-                                  "get-channels/processors/0",
-                                  "get-channels/processors/0/0",
-                                  "get-channels/processors/0/1",
-                                  "get-channels/processors/1",
-
-                                  "sample-config",
-                                  //
-                                  // "sample-config/0",
-                                  //
-                                  // "sample-config/1",
-                                  // "sample-config/1/connection",
-                                  // "sample-config/1/connection/0",
-                                  // "sample-config/1/connection/0/0",
-                                  //
-                                  "GetChannels",
-                                  "GetChannels/source",
-                                  "GetChannels/source/0",
-                                  "GetChannels/source/0/0",
-                                  "GetChannels/processors/0",
-
-                                  "_defaultGlobalElements",
-                                  //
-                                  // "_defaultGlobalElements/0",
-                                  // "_defaultGlobalElements/0/connection",
-                                  //
-                                  "RequestWithNoConfig",
-                                  "RequestWithNoConfig/processors/0",
-                                  "localhost-config-module-using-http-noconfig-default-config-global-element-suffix",
-                                  "localhost-config-module-using-http-noconfig-default-config-global-element-suffix/connection",
-                                  "do-request/processors/0",
-                                  "do-request/processors/1",
-
                                   "null",
                                   "null/0",
                                   "listenerConfig",
@@ -419,32 +378,7 @@ public class LazyInitConfigurationComponentLocatorTestCase extends AbstractInteg
 
                                   "os-config",
                                   "os-contains-flow",
-                                  "os-contains-flow/processors/0",
-
-                                  "request-with-oauth-auth-code-config-scConfig",
-                                  "request-with-oauth-auth-code-config-scConfig/connection",
-                                  "request-with-oauth-auth-code-config-scConfig/connection/0",
-                                  "request-with-oauth-auth-code-config-scConfig/connection/0/0",
-                                  "request-with-oauth-auth-code/processors/0",
-                                  "tokenManagerConfig-scConfig",
-                                  "listenerConfigOac",
-                                  "listenerConfigOac/connection",
-                                  "listen",
-                                  "listen/source",
-                                  "listen/processors/0",
-                                  "scConfig",
-                                  //
-                                  //"scConfig/0",
-                                  //
-                                  // "scConfig/1",
-                                  // "scConfig/1/connection",
-                                  // "scConfig/1/connection/0",
-                                  // "scConfig/1/connection/0/0",
-                                  //
-                                  "requestConfigOac",
-                                  "requestConfigOac/connection",
-                                  "request",
-                                  "request/processors/0"));
+                                  "os-contains-flow/processors/0"));
     assertThat(locator.find(builder().globalName("myFlow").build()), is(empty()));
     assertThat(locator.find(builder().globalName("anotherFlow").build()), is(empty()));
   }
@@ -649,25 +583,6 @@ public class LazyInitConfigurationComponentLocatorTestCase extends AbstractInteg
     assertThat(firstObj, not(sameInstance(secondObj)));
   }
 
-  @Test
-  public void lazyMuleContextSmartConnectorsWithConfig() throws IllegalAccessException {
-    lazyComponentInitializer.initializeComponents(componentLocation -> componentLocation.getLocation().equals("GetChannels"));
-
-    assertThat(locator.find(builderFromStringRepresentation("GetChannels").build()), is(not(empty())));
-    assertThat(locator.find(builderFromStringRepresentation("GetChannels/processors/0").build()), is(not(empty())));
-  }
-
-  @Test
-  @Issue("MULE-18259")
-  @Description("Default values for parameters declared in XML SDK configs were only working when XSD validations was enabled")
-  public void lazyMuleContextSmartConnectorsWithConfigAndDefaultParameters() throws IllegalAccessException {
-    lazyComponentInitializer.initializeComponents(componentLocation -> componentLocation.getLocation().equals("request"));
-
-    assertThat(locator.find(builderFromStringRepresentation("request").build()), is(not(empty())));
-    assertThat(locator.find(builderFromStringRepresentation("request/processors/0").build()), is(not(empty())));
-    assertThat(locator.find(builderFromStringRepresentation("listenerConfigOac").build()), is(not(empty())));
-  }
-
   @Description("Lazy init should create components that are references by other components, when the reference is not a top level element")
   @Test
   public void componentModelReferencesToNonTopLevelElement() {
@@ -762,14 +677,6 @@ public class LazyInitConfigurationComponentLocatorTestCase extends AbstractInteg
     lazyComponentInitializer.initializeComponent(builder().globalName("redeliveryPolicyFlowRef2").build());
 
     assertThat(locator.find(builder().globalName("redeliveryPolicyFlow").build()), is(not(empty())));
-  }
-
-  @Test
-  @Issue("MULE-18197")
-  @Description("Apps with Smart connectors wih default global elelements are properly initialized with LazyInit")
-  public void xmlSdkOperationWithDefaultConfig() {
-    lazyComponentInitializer.initializeComponent(builder().globalName("RequestWithNoConfig").build());
-    assertThat(locator.find(builder().globalName("RequestWithNoConfig").build()), is(not(empty())));
   }
 
 }
